06.16.2022 방로고 & 아이디 찾기

+++ GongGalBBang/GongGalBBang/helpers/RoomResponse.swift
@@ -19,5 +19,5 @@ struct RoomResult: Codable, Hashable {
 
 struct Datee: Codable, Hashable {
     var date : String
-    var member : Float
+    var member : Double
 }
